Converted to android-cmake for APK creation * Various changes to allow for Android to compile successfully

+++ src/CraftingRecipes.cpp
@@ -18,7 +18,7 @@
 cCraftingGrid::cCraftingGrid(int a_Width, int a_Height) :
 	m_Width(a_Width),
 	m_Height(a_Height),
-	m_Items(new cItem[a_Width * a_Height])
+	m_Items(new cItem[static_cast<size_t>(a_Width * a_Height)])
 {
 }
 
@@ -29,7 +29,7 @@ cCraftingGrid::cCraftingGrid(int a_Width, int a_Height) :
 cCraftingGrid::cCraftingGrid(const cItem * a_Items, int a_Width, int a_Height) :
 	m_Width(a_Width),
 	m_Height(a_Height),
-	m_Items(new cItem[a_Width * a_Height])
+	m_Items(new cItem[static_cast<size_t>(a_Width * a_Height)])
 {
 	for (int i = a_Width * a_Height - 1; i >= 0; i--)
 	{
@@ -44,7 +44,7 @@ cCraftingGrid::cCraftingGrid(const cItem * a_Items, int a_Width, int a_Height) :
 cCraftingGrid::cCraftingGrid(const cCraftingGrid & a_Original) :
 	m_Width(a_Original.m_Width),
 	m_Height(a_Original.m_Height),
-	m_Items(new cItem[a_Original.m_Width * a_Original.m_Height])
+	m_Items(new cItem[static_cast<size_t>(a_Original.m_Width * a_Original.m_Height)])
 {
 	for (int i = m_Width * m_Height - 1; i >= 0; i--)
 	{
